Added remote.py, Updated wsServer, Added writeWs
writeWs takes an object, converts it to json, and sends writes it via the wsServer.
Added write_battery_voltage, sends the battery voltage via writeWs.
Updated wsServer.cpp to accept remote input and writes it to the stdout prepended with a UID
Created remote.py, running this will allow the webpage to control parts of the robot (right now just the arm). It will write a pin to stdout, use this on the webpage to initiate control.